% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_approximation.R
\name{compute_moment}
\alias{compute_moment}
\title{Moment computation for the negative binomial posteror}
\usage{
compute_moment(
  k,
  data,
  ell_max = 50,
  a = 1,
  b = max(2, k + 1),
  c = 0,
  d = 1,
  z1 = 0,
  z2 = -1
)
}
\arguments{
\item{k}{Integer: The moment order, \eqn{\mathbb{E}[X^k]}}

\item{data}{Array: The count data with which to estimate the moment}

\item{ell_max}{Integer, the maximum degree of the polynomial approximation to the Gamma function}

\item{a, b}{the parameters for the Pearson-VI prior for the parameter m (odds ratio). Default
a=0, b=max(2, k+1). The prior has the form \eqn{m^a/(1+m)^b}.}

\item{c, d, z1, z2}{The parameters for the PEarson-VI prior for the parameter r
(dispersion parameter). Defaults c=0,d=1,z1=0,z2=-1. The prior takes the form
\eqn{(r-z1)^c/(r-z2)^d}. Note that the distribution is defined only for 0>=z1>z2>-sum(data), and
is supported on \eqn{(-z1, \infty)}. A minimally informative prior must set z1=0, and both z1 and z2 should be integers.}
}
\value{
The approximated moment, using the method in LeDuc and Kissler 2026.
}
\description{
Uses the polynomial approximation of Bradlow et al as described in LeDuc and Kissler 2026
to estimate moments of the negative binomial posterior distribution. Can compute both negative and positive integer moments
with conditions on the moments described in the paper.
}
