% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_approximation.R
\name{compute_moment_tricomi}
\alias{compute_moment_tricomi}
\title{Moment computation for the negative binomial posterior via the Tricomi expansion}
\usage{
compute_moment_tricomi(
  k,
  data,
  a = 1,
  b = max(2, k + 1),
  c = 0,
  d = 1,
  z1 = 0,
  z2 = -1
)
}
\arguments{
\item{k}{Integer: The moment order,\eqn{\mathbb{E}[X^k]}}

\item{data}{Array: The count data with which to estimate the moment}

\item{a, b}{the parameters for the Pearson-VI prior for the parameter m (odds ratio). Default
a=0, b=max(2, k+1). The prior has the form \eqn{m^a/(1+m)^b}.}

\item{c, d, z1, z2}{The parameters for the PEarson-VI prior for the parameter r
(dispersion parameter). Defaults c=0,d=1,z1=0,z2=-1. The prior takes the form
\eqn{(r-z1)^c/(r-z2)^d}. Note that the distribution is defined only for 0>=z1>z2>-sum(data), and
is supported on \eqn{(-z1, \infty)}. A minimally informative prior must set z1=0, and both z1 and z2 should be integers.}
}
\value{
Right now, an error. Eventually, the approximated moment using a method based on the Tricomi expansion.
}
\description{
Uses the Tricomi expansion
to estimate moments of the negative binomial posterior distribution. Still a
work in progress, but hopefully will allow generalization to arbitrary real moments
}
