% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_approximation.R
\name{compute_U_polynomial}
\alias{compute_U_polynomial}
\title{Computation of the polynomial approximant to the Gamma function.}
\usage{
compute_U_polynomial(k, x_sum, a, ell_max)
}
\arguments{
\item{k}{The moment to be approximated, \eqn{\mathbb{E}[X^k]}}

\item{x_sum}{the sum of the data}

\item{a}{The a parameter for the prior distribution of m}

\item{ell_max}{the maximum degree of the polynomial approximation}
}
\value{
The coefficients of the polynomial U as well as the log of the scaling
}
\description{
Computes the polynomial approximation to the Gamma function used in the
moment estimation. Does so in a relatively stable manner by normalizing the approximation
at each step.
}
