% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_measures.R
\name{credible_interval}
\alias{credible_interval}
\title{Calculate equal-tail credible interval}
\usage{
credible_interval(object, prob = 0.95, grid = NA)
}
\arguments{
\item{object}{maxent_fit object}

\item{prob}{probability mass (default 0.95)}
}
\value{
numeric vectors with lower and upper bounds
}
\description{
Compute equal-tail credible interval from MaxEnt density
}
