% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_measures.R
\name{dgig}
\alias{dgig}
\title{GIG PDF}
\usage{
dgig(x, p, a, b)
}
\arguments{
\item{x}{numeric vector of evaluation points}

\item{p}{GIG parameter p}

\item{a}{GIG parameter a (> 0)}

\item{b}{GIG parameter b (> 0)}
}
\value{
numeric vector of density values
}
\description{
Evaluate GIG density
}
