% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_measures.R
\name{dual_hessian}
\alias{dual_hessian}
\title{Hessian of the dual problem}
\usage{
dual_hessian(lambda, powers, nodes, weights, Xpow)
}
\arguments{
\item{lambda}{numeric vector of Lagrange multipliers}

\item{powers}{integer vector of moment orders}

\item{nodes}{quadrature nodes}

\item{weights}{quadrature weights}
}
\value{
matrix of second derivatives
}
\description{
Hessian of the dual objective (for diagnostics / Newton steps)
\eqn{H_{ij} \approx Cov_p[X^{k_i}, X^{k_j}]} in the Laplace approximation
}
