% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_measures.R
\name{estimate_dist}
\alias{estimate_dist}
\title{Model run function}
\usage{
estimate_dist(
  x1,
  x2,
  moments = c(-1, 1),
  ellmax = 120,
  n_nodes = 200,
  n_outer = 50,
  a = 1,
  b = max(2, max(moments) + 1),
  c = 0,
  d = 1,
  z1 = 0,
  z2 = -1
)
}
\arguments{
\item{x1}{Dataset corresponding to the numerator of the mean ratio}

\item{x2}{Dataset corresponding to the denominator of the mean ratio}

\item{moments}{The desired moments to be estimated, input as c(k1,k2,...,kn) for moments E{X^{k1}}, etc. Default is c(-1,1)}

\item{ellmax}{Maximum order of the polynomial approximation to the gamma function. Default 120}

\item{n_nodes}{Number of nodes to use in quadrature for moment evaluation. Default 200.}

\item{n_outer}{Number of outer iterations used to solve the moment problem for a given set of moments. Default 50.}

\item{a, b}{the parameters for the Pearson-VI prior for the parameter m (odds ratio). Default
a=0, b=max(2, k+1). The prior has the form \eqn{m^a/(1+m)^b}.}

\item{c, d, z1, z2}{The parameters for the PEarson-VI prior for the parameter r
(dispersion parameter). Defaults c=0,d=1,z1=0,z2=-1. The prior takes the form
\eqn{(r-z1)^c/(r-z2)^d}. Note that the distribution is defined only for 0>=z1>z2>-sum(data), and
is supported on \eqn{(-z1, \infty)}. A minimally informative prior must set z1=0, and both z1 and z2 should be integers.}
}
\value{
A maxent_fit object corresponding to the maximum entropy solution to the induced moment problem
}
\description{
Wrapper for the package, generates a maxent_fit object for a given dataset and
list of desired moments.
}
