% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_approximation.R
\name{fft_convolve}
\alias{fft_convolve}
\title{Convolution of two vectors via FFT}
\usage{
fft_convolve(a, b)
}
\arguments{
\item{a, b}{the vectors to convolve}
}
\value{
The convolution
}
\description{
Uses the FFT to convolve two vectors
}
