% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_measures.R
\name{fit_tnorm}
\alias{fit_tnorm}
\title{Fit the truncated normal distribution}
\usage{
fit_tnorm(mn, sqmn)
}
\arguments{
\item{mn}{Desired mean of the distribution}

\item{sqmn}{Desired value of \eqn{E[X^2]}}
}
\value{
vector of lagrange multipliers such that the posterior distribution is a truncated normal.
}
\description{
Solve for the maximum entropy measure on (0, inf) given a mean and second moment

Given a target mean and second moment, calculates the truncated normal distribution that maximizes the entropy given those constraints.
}
