% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_measures.R
\name{gauss_laguerre_quad}
\alias{gauss_laguerre_quad}
\title{Gauss-Laguerre based quadrature rule}
\usage{
gauss_laguerre_quad(lambda, moments, n_nodes = 100)
}
\arguments{
\item{lambda}{The Lagrange multipliers for the moments}

\item{n_nodes}{Number of quadrature nodes}

\item{powers}{Integer vector of moment orders}
}
\value{
A list describing the quadrature rule
}
\description{
Quadrature rule based on Gauss-Laguerre quadrature for estimating the moments
}
