% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_measures.R
\name{log_density_nodes}
\alias{log_density_nodes}
\title{Unnormalized log density}
\usage{
log_density_nodes(lambda, Xpow)
}
\arguments{
\item{lambda}{numeric vector of Lagrange multipliers}

\item{powers}{integer vector of moment orders}

\item{nodes}{numeric vector of quadrature nodes}
}
\value{
numeric vector of log density values at nodes
}
\description{
Evaluate log unnormalized density at quadrature nodes
}
