% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_measures.R
\name{normalized_weights}
\alias{normalized_weights}
\title{Calculates normalized weights so that \eqn{\int fdx=1}}
\usage{
normalized_weights(lambda, powers, nodes, weights, Xpow)
}
\arguments{
\item{lambda}{numeric vector of Lagrange multipliers}

\item{powers}{integer vector of moment orders}

\item{nodes}{numeric vector of quadrature nodes}

\item{weights}{numeric vector of quadrature weights}
}
\value{
list with log_Z and normalized probability weights at nodes
}
\description{
Compute normalized weights at quadrature nodes
}
