% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_measures.R
\name{plot.maxent_fit}
\alias{plot.maxent_fit}
\title{Plot the MaxEnt density}
\usage{
\method{plot}{maxent_fit}(x, n_plot = 500, log_scale = FALSE, add = FALSE, ...)
}
\arguments{
\item{x}{maxent_fit object}

\item{n_plot}{number of points for smooth curve}

\item{log_scale}{plot on log x scale (default TRUE)}

\item{add}{add to existing plot (default FALSE)}

\item{...}{additional arguments passed to plot/lines}
}
\description{
S3 method for plotting the maxent density
}
