% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_measures.R
\name{predict.maxent_fit}
\alias{predict.maxent_fit}
\title{Predict method for maxent_fit}
\usage{
\method{predict}{maxent_fit}(object, newdata, ...)
}
\arguments{
\item{object}{maxent_fit object}

\item{newdata}{numeric vector of x values}
}
\value{
numeric vector of normalized density values
}
\description{
Evaluate MaxEnt density at new points
}
